requirements.txt
django==5.0.1
django_ninja==1.1.0
django-redis==5.4.0
celery==5.3.6
requests==2.25.1
ipdb==0.13.4                 # BSD
